[Layer] Update layers (#249)
