+++ cogdl/layers/__init__.py
@@ -1,5 +1,36 @@
-from .maggregator import MeanAggregator, SumAggregator
+from .gcn_layer import GCNLayer
+from .sage_layer import MeanAggregator, SumAggregator, SAGELayer
+from .gat_layer import GATLayer
+from .gin_layer import GINLayer
 from .se_layer import SELayer
+from .deepergcn_layer import GENConv, DeepGCNLayer
+from .disengcn_layer import DisenGCNLayer
+from .gcnii_layer import GCNIILayer
+from .mlp_layer import MLPLayer
+from .saint_layer import SAINTLayer
+from .han_layer import HANLayer
+from .pprgo_layer import PPRGoLayer
+from .rgcn_layer import RGCNLayer
+from .sgc_layer import SGCLayer
 from .mixhop_layer import MixHopLayer
 
-__all__ = ["SELayer", "MeanAggregator", "SumAggregator", "MixHopLayer"]
+__all__ = [
+    "GCNLayer",
+    "MeanAggregator",
+    "SumAggregator",
+    "SAGELayer",
+    "GATLayer",
+    "GINLayer",
+    "SELayer",
+    "GENConv",
+    "DeepGCNLayer",
+    "DisenGCNLayer",
+    "GCNIILayer",
+    "MLPLayer",
+    "SAINTLayer",
+    "HANLayer",
+    "PPRGoLayer",
+    "RGCNLayer",
+    "SGCLayer",
+    "MixHopLayer",
+]
