+++ cogdl/layers/mlp_layer.py
@@ -1,0 +1,35 @@
+import math
+
+import torch
+import torch.nn as nn
+import torch.nn.functional as F
+
+
+class MLPLayer(nn.Module):
+    def __init__(self, in_features, out_features, bias=True):
+        super(MLPLayer, self).__init__()
+        self.in_features = in_features
+        self.out_features = out_features
+        self.weight = nn.Parameter(torch.FloatTensor(in_features, out_features))
+        if bias:
+            self.bias = nn.Parameter(torch.FloatTensor(out_features))
+        else:
+            self.register_parameter("bias", None)
+        self.reset_parameters()
+
+    def reset_parameters(self):
+
+        stdv = 1.0 / math.sqrt(self.weight.size(1))
+        self.weight.data.normal_(-stdv, stdv)
+        if self.bias is not None:
+            self.bias.data.normal_(-stdv, stdv)
+
+    def forward(self, x):
+        output = torch.mm(x, self.weight)
+        if self.bias is not None:
+            return output + self.bias
+        else:
+            return output
+
+    def __repr__(self):
+        return self.__class__.__name__ + " (" + str(self.in_features) + " -> " + str(self.out_features) + ")"
