+++ cogdl/models/nn/drgcn.py
@@ -2,10 +2,9 @@ import numpy as np
 import torch.nn as nn
 import torch.nn.functional as F
 
-from cogdl.layers import SELayer
+from cogdl.layers import SELayer, GCNLayer
 
 from .. import BaseModel, register_model
-from .gcn import GraphConvolution
 
 
 @register_model("drgcn")
@@ -40,7 +39,7 @@ class DrGCN(BaseModel):
         self.num_layers = num_layers
         self.dropout = dropout
         shapes = [num_features] + [hidden_size] * (num_layers - 1) + [num_classes]
-        self.convs = nn.ModuleList([GraphConvolution(shapes[layer], shapes[layer + 1]) for layer in range(num_layers)])
+        self.convs = nn.ModuleList([GCNLayer(shapes[layer], shapes[layer + 1]) for layer in range(num_layers)])
         self.ses = nn.ModuleList(
             [SELayer(shapes[layer], se_channels=int(np.sqrt(shapes[layer]))) for layer in range(num_layers)]
         )
