+++ cogdl/models/nn/gat.py
@@ -1,94 +1,8 @@
-import torch
 import torch.nn as nn
 import torch.nn.functional as F
-import math
 
+from cogdl.layers import GATLayer
 from .. import BaseModel, register_model
-from cogdl.utils import mul_edge_softmax, spmm, mh_spmm, check_mh_spmm
-
-
-class GATLayer(nn.Module):
-    """
-    Sparse version GAT layer, similar to https://arxiv.org/abs/1710.10903
-    """
-
-    def __init__(self, in_features, out_features, nhead=1, alpha=0.2, dropout=0.6, concat=True, residual=False):
-        super(GATLayer, self).__init__()
-        self.in_features = in_features
-        self.out_features = out_features
-        self.alpha = alpha
-        self.concat = concat
-        self.nhead = nhead
-
-        self.W = nn.Parameter(torch.FloatTensor(in_features, out_features * nhead))
-
-        self.a_l = nn.Parameter(torch.zeros(size=(1, nhead, out_features)))
-        self.a_r = nn.Parameter(torch.zeros(size=(1, nhead, out_features)))
-
-        self.dropout = nn.Dropout(dropout)
-        self.leakyrelu = nn.LeakyReLU(self.alpha)
-
-        if residual:
-            out_features = out_features * nhead if concat else out_features
-            self.residual = nn.Linear(in_features, out_features)
-        else:
-            self.register_buffer("residual", None)
-        self.reset_parameters()
-
-    def reset_parameters(self):
-        def reset(tensor):
-            stdv = math.sqrt(6.0 / (tensor.size(-2) + tensor.size(-1)))
-            tensor.data.uniform_(-stdv, stdv)
-
-        reset(self.a_l)
-        reset(self.a_r)
-        reset(self.W)
-
-        # nn.init.xavier_uniform_(self.W.data, gain=1.414)
-        # nn.init.xavier_uniform_(self.a_r.data, gain=1.414)
-        # nn.init.xavier_uniform_(self.a_l.data, gain=1.414)
-
-    def forward(self, graph, x):
-        h = torch.matmul(x, self.W).view(-1, self.nhead, self.out_features)
-        h[torch.isnan(h)] = 0.0
-
-        row, col = graph.edge_index
-        # Self-attention on the nodes - Shared attention mechanism
-        h_l = (self.a_l * h).sum(dim=-1)[row]
-        h_r = (self.a_r * h).sum(dim=-1)[col]
-        edge_attention = self.leakyrelu(h_l + h_r)
-        # edge_attention: E * H
-        edge_attention = mul_edge_softmax(graph, edge_attention)
-        edge_attention = self.dropout(edge_attention)
-
-        if check_mh_spmm() and next(self.parameters()).device.type != "cpu":
-            if self.nhead > 1:
-                h_prime = mh_spmm(graph, edge_attention, h)
-                out = h_prime.view(h_prime.shape[0], -1)
-            else:
-                edge_weight = edge_attention.view(-1)
-                with graph.local_graph():
-                    graph.edge_weight = edge_weight
-                    out = spmm(graph, h.squeeze(1))
-        else:
-            with graph.local_graph():
-                h_prime = []
-                h = h.permute(1, 0, 2).contiguous()
-                for i in range(self.nhead):
-                    edge_weight = edge_attention[:, i]
-                    graph.edge_weight = edge_weight
-                    hidden = h[i]
-                    assert not torch.isnan(hidden).any()
-                    h_prime.append(spmm(graph, hidden))
-            out = torch.cat(h_prime, dim=1)
-
-        if self.residual:
-            res = self.residual(x)
-            out += res
-        return out
-
-    def __repr__(self):
-        return self.__class__.__name__ + " (" + str(self.in_features) + " -> " + str(self.out_features) + ")"
 
 
 @register_model("gat")
