+++ cogdl/models/nn/gcnii.py
@@ -4,32 +4,8 @@ import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
+from cogdl.layers import GCNIILayer
 from .. import register_model, BaseModel
-from cogdl.utils import symmetric_normalization, add_remaining_self_loops, spmm
-
-
-class GCNIILayer(nn.Module):
-    def __init__(self, n_channels, alpha=0.1, beta=1, residual=False):
-        super(GCNIILayer, self).__init__()
-        self.n_channels = n_channels
-        self.alpha = alpha
-        self.beta = beta
-        self.residual = residual
-        self.weight = nn.Parameter(torch.FloatTensor(n_channels, n_channels))
-        self.reset_parameters()
-
-    def reset_parameters(self):
-        stdv = 1.0 / math.sqrt(self.n_channels)
-        self.weight.data.uniform_(-stdv, stdv)
-
-    def forward(self, graph, x, init_x):
-        """Symmetric normalization"""
-        hidden = spmm(graph, x)
-        hidden = (1 - self.alpha) * hidden + self.alpha * init_x
-        h = self.beta * torch.matmul(hidden, self.weight) + (1 - self.beta) * hidden
-        if self.residual:
-            h = h + x
-        return h
 
 
 @register_model("gcnii")
