+++ cogdl/models/nn/graphsaint.py
@@ -4,116 +4,8 @@ import torch.nn.functional as F
 import numpy as np
 
 from .. import BaseModel, register_model
+from cogdl.layers import SAINTLayer
 from cogdl.trainers.sampled_trainer import SAINTTrainer
-from cogdl.utils import spmm
-
-
-F_ACT = {"relu": nn.ReLU(), "I": lambda x: x}
-
-"""
-Borrowed from https://github.com/GraphSAINT/GraphSAINT
-"""
-
-
-class HighOrderAggregator(nn.Module):
-    def __init__(self, dim_in, dim_out, dropout=0.0, act="relu", order=1, aggr="mean", bias="norm-nn", **kwargs):
-        """
-        Layer implemented here combines the GraphSAGE-mean [1] layer with MixHop [2] layer.
-        We define the concept of `order`: an order-k layer aggregates neighbor information
-        from 0-hop all the way to k-hop. The operation is approximately:
-            X W_0 [+] A X W_1 [+] ... [+] A^k X W_k
-        where [+] is some aggregation operation such as addition or concatenation.
-
-        Special cases:
-            Order = 0  -->  standard MLP layer
-            Order = 1  -->  standard GraphSAGE layer
-
-        [1]: https://cs.stanford.edu/people/jure/pubs/graphsage-nips17.pdf
-        [2]: https://arxiv.org/abs/1905.00067
-
-        Inputs:
-            dim_in      int, feature dimension for input nodes
-            dim_out     int, feature dimension for output nodes
-            dropout     float, dropout on weight matrices W_0 to W_k
-            act         str, activation function. See F_ACT at the top of this file
-            order       int, see definition above
-            aggr        str, if 'mean' then [+] operation adds features of various hops
-                            if 'concat' then [+] concatenates features of various hops
-            bias        str, if 'bias' then apply a bias vector to features of each hop
-                            if 'norm' then perform batch-normalization on output features
-
-        Outputs:
-            None
-        """
-        super(HighOrderAggregator, self).__init__()
-        assert bias in ["bias", "norm", "norm-nn"]
-        self.order, self.aggr = order, aggr
-        self.act, self.bias = F_ACT[act], bias
-        self.dropout = dropout
-        self.f_lin, self.f_bias = [], []
-        self.offset, self.scale = [], []
-        self.num_param = 0
-        for o in range(self.order + 1):
-            self.f_lin.append(nn.Linear(dim_in, dim_out, bias=False))
-            nn.init.xavier_uniform_(self.f_lin[-1].weight)
-            self.f_bias.append(nn.Parameter(torch.zeros(dim_out)))
-            self.num_param += dim_in * dim_out
-            self.num_param += dim_out
-            self.offset.append(nn.Parameter(torch.zeros(dim_out)))
-            self.scale.append(nn.Parameter(torch.ones(dim_out)))
-            if self.bias == "norm" or self.bias == "norm-nn":
-                self.num_param += 2 * dim_out
-        self.f_lin = nn.ModuleList(self.f_lin)
-        self.f_dropout = nn.Dropout(p=self.dropout)
-        self.params = nn.ParameterList(self.f_bias + self.offset + self.scale)
-        self.f_bias = self.params[: self.order + 1]
-        if self.bias == "norm":
-            self.offset = self.params[self.order + 1 : 2 * self.order + 2]
-            self.scale = self.params[2 * self.order + 2 :]
-        elif self.bias == "norm-nn":
-            final_dim_out = dim_out * ((aggr == "concat") * (order + 1) + (aggr == "mean"))
-            self.f_norm = nn.BatchNorm1d(final_dim_out, eps=1e-9, track_running_stats=True)
-        self.num_param = int(self.num_param)
-
-    def _f_feat_trans(self, _feat, _id):
-        feat = self.act(self.f_lin[_id](_feat) + self.f_bias[_id])
-        if self.bias == "norm":
-            mean = feat.mean(dim=1).view(feat.shape[0], 1)
-            var = feat.var(dim=1, unbiased=False).view(feat.shape[0], 1) + 1e-9
-            feat_out = (feat - mean) * self.scale[_id] * torch.rsqrt(var) + self.offset[_id]
-        else:
-            feat_out = feat
-        return feat_out
-
-    def forward(self, input):
-        """
-        Inputs:.
-            adj_norm        normalized adj matrix of the subgraph
-            feat_in         2D matrix of input node features
-
-        Outputs:
-            adj_norm        same as input (to facilitate nn.Sequential)
-            feat_out        2D matrix of output node features
-        """
-
-        graph, x = input
-        feat_in = self.f_dropout(x)
-        feat_hop = [feat_in]
-        # generate A^i X
-        for o in range(self.order):
-            feat_hop.append(spmm(graph, x))
-        feat_partial = [self._f_feat_trans(ft, idf) for idf, ft in enumerate(feat_hop)]
-        if self.aggr == "mean":
-            feat_out = feat_partial[0]
-            for o in range(len(feat_partial) - 1):
-                feat_out += feat_partial[o + 1]
-        elif self.aggr == "concat":
-            feat_out = torch.cat(feat_partial, 1)
-        else:
-            raise NotImplementedError
-        if self.bias == "norm-nn":
-            feat_out = self.f_norm(feat_out)
-        return graph, feat_out  # return adj_norm to support Sequential
 
 
 def parse_arch(architecture, aggr, act, bias, hidden_size, num_features):
@@ -176,7 +68,7 @@ class GraphSAINT(BaseModel):
             None
         """
         super(GraphSAINT, self).__init__()
-        self.aggregator_cls = HighOrderAggregator
+        self.aggregator_cls = SAINTLayer
         self.mulhead = 1
         self.weight_decay = weight_decay
         self.dropout = dropout
@@ -197,8 +89,8 @@ class GraphSAINT(BaseModel):
         self.num_params = 0
         self.aggregators, num_param = self.get_aggregators()
         self.num_params += num_param
-        self.conv_layers = nn.Sequential(*self.aggregators)
-        self.classifier = HighOrderAggregator(
+        self.conv_layers = nn.ModuleList(self.aggregators)
+        self.classifier = SAINTLayer(
             self.dims_feat[-1], self.num_classes, act="I", order=0, dropout=self.dropout, bias="bias"
         )
         self.num_params += self.classifier.num_param
@@ -237,9 +129,10 @@ class GraphSAINT(BaseModel):
 
     def forward(self, graph):
         x = graph.x
-        _, emb_subg = self.conv_layers(((graph, x)))
-        emb_subg_norm = F.normalize(emb_subg, p=2, dim=1)
-        pred_subg = self.classifier((None, emb_subg_norm))[1]
+        for layer in self.conv_layers:
+            x = layer(graph, x)
+        emb_subg_norm = F.normalize(x, p=2, dim=1)
+        pred_subg = self.classifier(None, emb_subg_norm)
         return pred_subg
 
     def _loss(self, preds, labels, norm_loss):
@@ -275,7 +168,6 @@ class GraphSAINT(BaseModel):
 
     def predict(self, data):
         return self.forward(data)
-        # return nn.Sigmoid()(preds) if self.sigmoid_loss else F.softmax(preds, dim=1
 
     @staticmethod
     def get_trainer(task, args):
