+++ cogdl/models/nn/sgc.py
@@ -1,22 +1,6 @@
-import torch.nn as nn
+from cogdl.layers import SGCLayer
 
 from .. import BaseModel, register_model
-from cogdl.utils import spmm
-
-
-class SimpleGraphConvolution(nn.Module):
-    def __init__(self, in_features, out_features, order=3):
-        super(SimpleGraphConvolution, self).__init__()
-        self.in_features = in_features
-        self.out_features = out_features
-        self.order = order
-        self.W = nn.Linear(in_features, out_features)
-
-    def forward(self, graph, x):
-        output = self.W(x)
-        for _ in range(self.order):
-            output = spmm(graph, output)
-        return output
 
 
 @register_model("sgc")
@@ -32,7 +16,7 @@ class sgc(BaseModel):
 
     def __init__(self, in_feats, out_feats):
         super(sgc, self).__init__()
-        self.nn = SimpleGraphConvolution(in_feats, out_feats)
+        self.nn = SGCLayer(in_feats, out_feats)
         self.cache = dict()
 
     def forward(self, graph):
