+++ cogdl/models/nn/sortpool.py
@@ -1,13 +1,11 @@
-import random
-
 import torch
 import torch.nn as nn
 import torch.nn.functional as F
 
+from cogdl.layers import SAGELayer
+from cogdl.utils import split_dataset_general
+
 from .. import BaseModel, register_model
-from cogdl.data import DataLoader
-from .graphsage import GraphSAGELayer
-from .gin import split_dataset_general
 
 
 def scatter_sum(src, index, dim, dim_size):
@@ -98,9 +96,9 @@ class SortPool(BaseModel):
         self.dropout = dropout
         self.num_layers = num_layers
         self.gnn_convs = nn.ModuleList()
-        self.gnn_convs.append(GraphSAGELayer(in_feats, hidden_dim))
+        self.gnn_convs.append(SAGELayer(in_feats, hidden_dim))
         for _ in range(self.num_layers - 1):
-            self.gnn_convs.append(GraphSAGELayer(hidden_dim, hidden_dim))
+            self.gnn_convs.append(SAGELayer(hidden_dim, hidden_dim))
         self.conv1d = nn.Conv1d(hidden_dim, out_channel, kernel_size)
         self.fc1 = nn.Linear(out_channel * (self.k - kernel_size + 1), hidden_dim)
         self.fc2 = nn.Linear(hidden_dim, num_classes)
