+++ docs/requirements.txt
@@ -17,11 +17,11 @@ ogb
 black
 pytest
 coveralls
-https://download.pytorch.org/whl/cpu/torch-1.6.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
-https://pytorch-geometric.com/whl/torch-1.6.0+cpu/torch_scatter-2.0.5-cp36-cp36m-linux_x86_64.whl
-https://pytorch-geometric.com/whl/torch-1.6.0+cpu/torch_sparse-0.6.8-cp36-cp36m-linux_x86_64.whl
-https://pytorch-geometric.com/whl/torch-1.6.0+cpu/torch_cluster-1.5.8-cp36-cp36m-linux_x86_64.whl
-https://pytorch-geometric.com/whl/torch-1.6.0+cpu/torch_spline_conv-1.2.0-cp36-cp36m-linux_x86_64.whl
+https://download.pytorch.org/whl/cpu/torch-1.7.1%2Bcpu-cp36-cp36m-linux_x86_64.whl
+https://pytorch-geometric.com/whl/torch-1.7.0+cpu/torch_scatter-2.0.7-cp36-cp36m-linux_x86_64.whl
+https://pytorch-geometric.com/whl/torch-1.7.0+cpu/torch_sparse-0.6.9-cp36-cp36m-linux_x86_64.whl
+https://pytorch-geometric.com/whl/torch-1.7.0+cpu/torch_cluster-1.5.9-cp36-cp36m-linux_x86_64.whl
+https://pytorch-geometric.com/whl/torch-1.7.0+cpu/torch_spline_conv-1.2.1-cp36-cp36m-linux_x86_64.whl
 torch-geometric
 dgl==0.4.3
 numba
