+++ docs/source/api/layers.rst
@@ -1,6 +1,84 @@
 layers
 ======
 
+Layers
+-------------------------
+
+.. automodule:: cogdl.layers.gcn_layer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: cogdl.layers.gat_layer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: cogdl.layers.sage_layer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: cogdl.layers.gin_layer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: cogdl.layers.gcnii_layer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: cogdl.layers.deepergcn_layer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: cogdl.layers.disengcn_layer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: cogdl.layers.han_layer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: cogdl.layers.mlp_layer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: cogdl.layers.pprgo_layer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: cogdl.layers.rgcn_layer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: cogdl.layers.saint_layer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: cogdl.layers.sgc_layer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: cogdl.layers.mixhop_layer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+.. automodule:: cogdl.layers.se_layer
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
 GCC module
 -------------------------
 
@@ -25,22 +103,6 @@ Link Prediction module
     :undoc-members:
     :show-inheritance:
 
-Mean Aggregator module
--------------------------
-
-.. automodule:: cogdl.layers.maggregator
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
-MixHop module
----------------------------
-
-.. automodule:: cogdl.layers.mixhop_layer
-    :members:
-    :undoc-members:
-    :show-inheritance:
-
 PPRGo module
 -------------------------
 
@@ -57,13 +119,6 @@ ProNE module
     :undoc-members:
     :show-inheritance:
 
-SELayer module
------------------------
-
-.. automodule:: cogdl.layers.se_layer
-    :members:
-    :undoc-members:
-    :show-inheritance:
 
 SRGCN module
 -------------------------
