+++ examples/oagbert/oagbert.py
@@ -1,9 +1,13 @@
+import torch
 from cogdl import oagbert
 
 tokenizer, bert_model = oagbert()
+bert_model.eval()
 
 sequence = ["CogDL is developed by KEG, Tsinghua.", "OAGBert is developed by KEG, Tsinghua."]
 tokens = tokenizer(sequence, return_tensors="pt", padding=True)
-outputs = bert_model(**tokens)
 
-print(outputs[0].shape)
+with torch.no_grad():
+    outputs = bert_model(**tokens)
+
+print(outputs[0])
